feat: add tag type and change placeId type
