+++ src/components/place/types.ts
@@ -1,5 +1,7 @@
+import { PlaceType } from '@/types/api/place'
+
 export interface PlaceProps {
-  placeId: number
+  placeId: PlaceType['place']['id']
   name: string
   address: string
   distance?: string
@@ -11,41 +13,4 @@ export interface PlaceProps {
     hashtags?: string[]
     onClickLike: () => void
   }
-}
-
-export interface PlaceType {
-  place: {
-    id: number
-    kakaoPlace: {
-      id: number
-      name: string
-      category: string
-      address: string
-      x: number
-      y: number
-      menuList: {
-        menu: string
-        price: string
-      }[]
-      photoList: string[]
-    }
-    x: number
-    y: number
-  }
-  tags: [
-    {
-      id: number
-      mapId: string
-      content: string
-      createdAt: string
-    },
-  ]
-  comments: [{}]
-  likedUserIds: number[]
-  createdBy: {
-    id: number
-    nickname: string
-  }
-  createdAt: string
-  updatedAt: string
 }
