+++ src/types/api/maps.ts
@@ -1,0 +1,6 @@
+export interface TagItem {
+  id: number
+  mapId: string
+  content: string
+  createdAt: string
+}
